LayoutTests/fast/regex/non-capturing-backtracking.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
description(
"This page tests for proper backtracking with greedy quantifiers and non-capturing parentheses."
);

var re = /(?:a*)a/;
shouldBe("re.exec('a')", "['a']");
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
