safari-534.53-branch/LayoutTests/fast/regex/non-capturing-backtracking-expected.txt

This page tests for proper backtracking with greedy quantifiers and non-capturing parentheses.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS re.exec('a') is ['a']
PASS successfullyParsed is true

TEST COMPLETE
